% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBPKapp.R
\name{PBPKapp}
\alias{PBPKapp}
\title{Run the PBPK Application}
\usage{
PBPKapp()
}
\value{
No return value.
}
\description{
This function runs the PBPK application. The user can alter certain properties
and then run the exposures with amounts reported on the graph.
}
\examples{
# Run the PBPK app.
if (interactive()) {
  PBPKapp()
}
}
